moved business logic from representation layer to core
